LayoutTests/accessibility/ax-thread-text-apis/text-marker-paragraph-nav-expected.txt

This tests that paragraph navigation is working correctly.

Current character is: p
Current paragraph is: paragraph test
Current character is: T
Current paragraph is: Test Content
Current character is: c
Current paragraph is: c [ATTACHMENT]d
Current character is: d
Current paragraph is: c [ATTACHMENT]d
Current character is: t
Current paragraph is: test audio [ATTACHMENT]file
Current character is: c
Current paragraph is: can't select
Current character is: 巧
Current paragraph is: 巧克力 是食物吗?
Current character is: ك
Current paragraph is: كيف حالك؟
Current character is: b
Current paragraph is: both   spaces
Current character is: i
Current paragraph is: line breaks
Current character is: s
Current paragraph is: some
Current character is: t
Current paragraph is: text

PASS successfullyParsed is true

TEST COMPLETE
paragraph test
Test Contenteditable is working.
c d
test audio file
can't select
巧克力 是食物吗?
كيف حالك؟
both   spaces
line breaks
some
text
this is my first paragraph. Of text. it has some text.


this is my second paragraph. Of text. it has some text.
this is my third paragraph. Of text. it has some text.
